# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TOAN Store E-commerce API
  description: |
    API documentation for TOAN Store E-commerce platform.
    
    **Authentication**: JWT via HTTP-Only Cookies (`nike_auth_session` / `nike_admin_session`)  
    **Rate Limiting**: Redis-backed sliding window, Fail-Closed for sensitive endpoints.
  version: 2.5.0
  contact:
    email: admin@toanstore.com
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful, sets cookies
          headers:
            Set-Cookie:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/account/update:
    put:
      tags: [Account]
      summary: Update user profile
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/cart/check-balance:
    post:
      tags: [Cart]
      summary: Check gift card balance (rate limited — 10 req/60s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cardNumber, pin]
              properties:
                cardNumber: { type: string, pattern: '^\d{16}$' }
                pin: { type: string, pattern: '^\d{4}$' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ═══════════════════════════════════
  # ORDERS
  # ═══════════════════════════════════
...
  /api/contact:
    post:
      tags: [Content]
      summary: Submit contact form
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nike_auth_session
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: nike_admin_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              message: { type: string, example: 'Too many requests. Please try again later.' }
              retryAfterSeconds: { type: integer }